changing to int

+++ SerialPrograms/Source/CommonFramework/Options/EditableTableOption.cpp
@@ -91,14 +91,14 @@ template<typename T> QWidget* make_integer_table_cell(QWidget& parent, T& value)
 }
 
 
-QWidget* make_limited_integer_table_cell(QWidget& parent, uint8_t& value, uint8_t min, uint8_t max){
+QWidget* make_limited_integer_table_cell(QWidget& parent, int& value, int min, int max){
     QLineEdit* box = new QLineEdit(QString::number(value), &parent);
     box->setAlignment(Qt::AlignHCenter);
     box->connect(
         box, &QLineEdit::textChanged,
         box, [&value, box, min, max](const QString& text){
             bool ok = false;
-            const uint8_t current = (uint8_t)text.toLong(&ok);
+            const int current = (int)text.toLong(&ok);
             QPalette palette;
             if (ok && current >= min && current <= max){
                 value = current;
